승아 - 바로구매 버튼 연결, 주문 리스트 삭제기능

+++ cosmetic/src/main/java/com/example/cosmetic/model/order/OrderDTO.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @ToString
 public class OrderDTO {
 	
-    private int orderid;
+    private long orderid;
     private String userid;
     
     private String orderItemId;
